django_project/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
django-extensions==3.1.3
Faker==8.12.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
